requirements.txt
streamlit
google-cloud-aiplatform
speechrecognition
google-cloud-storage
PyAudio
transformers
sentencepiece
torch
scipy